requirements.txt
streamlit
google-generativeai
python-pptx
pillow
pymupdf
streamlit-paste-button